# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sugestões de presentes personalizadas.",
        "title": "Gift App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/friends/{friendId}/profile": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Buscar perfil do amigo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do amigo",
                        "name": "friendId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Profile"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Exemplo de payload: {\"friendID\":\"9b02ce54-4f42-4a8b-a539-5b53a6e37e63\",\"likes\":[\"fotografia\",\"musica\",\"viagem\"],\"dislikes\":[\"multidoes\",\"atraso\"],\"personality\":[\"introvertida\",\"criativa\"],\"embedding\":[0.12,-0.44,0.91]}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profiles"
                ],
                "summary": "Criar ou atualizar perfil do amigo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do amigo",
                        "name": "friendId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do perfil",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ProfileUpsertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/profiles/agent/chat": {
            "post": {
                "description": "Exemplo de payload: {\"friendID\":\"9b02ce54-4f42-4a8b-a539-5b53a6e37e63\",\"message\":\"Ela prefere experiencias ou objetos?\"}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-agent"
                ],
                "summary": "Conversar com profile agent",
                "parameters": [
                    {
                        "description": "Payload do chat",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AgentChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/agent/finalize": {
            "post": {
                "description": "Exemplo de payload: {\"friendID\":\"9b02ce54-4f42-4a8b-a539-5b53a6e37e63\"}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-agent"
                ],
                "summary": "Finalizar sessão de profile agent",
                "parameters": [
                    {
                        "description": "Payload de finalização",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AgentFinalizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/agent/session/{friendId}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-agent"
                ],
                "summary": "Remover sessão do profile agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do friend",
                        "name": "friendId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Profile": {
            "type": "object",
            "properties": {
                "dislikes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "embedding": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "friendID": {
                    "type": "string"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "personality": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "http.AgentChatRequest": {
            "type": "object",
            "properties": {
                "friendID": {
                    "type": "string",
                    "example": "9b02ce54-4f42-4a8b-a539-5b53a6e37e63"
                },
                "message": {
                    "type": "string",
                    "example": "Ela prefere experiencias ou objetos?"
                }
            }
        },
        "http.AgentFinalizeRequest": {
            "type": "object",
            "properties": {
                "friendID": {
                    "type": "string",
                    "example": "9b02ce54-4f42-4a8b-a539-5b53a6e37e63"
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "birthDate must be in format YYYY-MM-DD"
                }
            }
        },
...
        "http.ProfileUpsertRequest": {
            "type": "object",
            "properties": {
                "dislikes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "multidoes",
                        "atraso"
                    ]
                },
                "embedding": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        0.12,
                        -0.44,
                        0.91
                    ]
                },
                "friendID": {
                    "type": "string",
                    "example": "9b02ce54-4f42-4a8b-a539-5b53a6e37e63"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "fotografia",
                        "musica",
                        "viagem"
                    ]
                },
                "personality": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "introvertida",
                        "criativa",
                        "detalhista"
                    ]
                }
            }
        },
...
    }
}